requirements.txt
Flask
Flask-Babel
Pillow
opencv-python
matplotlib